Activity Changes

+++ Adventureland-Village/src/main/java/com/adventurelandVillage/controller/ActivityController.java
@@ -47,7 +47,7 @@ public class ActivityController {
 
 	// Get activity by ID
 	@GetMapping("/{id}")
-	public ResponseEntity<Activity> getActivityById(@PathVariable(value = "id") int activityId)
+	public ResponseEntity<Activity> getActivityById(@PathVariable(value = "id") Long activityId)
 			throws ConfigDataResourceNotFoundException {
 		Activity activity = activityService.getActivityById(activityId);
 		return ResponseEntity.ok().body(activity);
@@ -55,7 +55,7 @@ public class ActivityController {
 
 	// Update activity
 	@PutMapping("/{id}")
-	public ResponseEntity<Activity> updateActivity(@PathVariable(value = "id") int activityId,
+	public ResponseEntity<Activity> updateActivity(@PathVariable(value = "id") Long activityId,
 			@Valid @RequestBody Activity activityDetails) throws ConfigDataResourceNotFoundException {
 		Activity updatedActivity = activityService.updateActivity(activityDetails);
 		return ResponseEntity.ok(updatedActivity);
